refactor: domain과 무관한 클래스 util로 분리
