+++ src/main/java/util/CsvReader.java
@@ -1,5 +1,6 @@
-package domain;
+package util;
 
+import domain.ErrorCode;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
