+++ src/main/java/util/Parser.java
@@ -1,5 +1,6 @@
-package domain;
+package util;
 
+import domain.ErrorCode;
 import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalTime;
